+++ RxCocoa/Traits/Signal/Signal+Subscription.swift
@@ -34,10 +34,32 @@ extension SharedSequenceConvertibleType where SharingStrategy == SignalSharingSt
     }
 
     /**
-     Creates new subscription and sends elements to variable.
+     Creates new subscription and sends elements to `BehaviorRelay`.
+     - parameter relay: Target relay for sequence elements.
+     - returns: Disposable object that can be used to unsubscribe the observer from the relay.
+     */
+    public func emit(to relay: BehaviorRelay<E>) -> Disposable {
+        return emit(onNext: { e in
+            relay.accept(e)
+        })
+    }
+    
+    /**
+     Creates new subscription and sends elements to `BehaviorRelay`.
+     - parameter relay: Target relay for sequence elements.
+     - returns: Disposable object that can be used to unsubscribe the observer from the relay.
+     */
+    public func emit(to relay: BehaviorRelay<E?>) -> Disposable {
+        return emit(onNext: { e in
+            relay.accept(e)
+        })
+    }
+    
+    /**
+     Creates new subscription and sends elements to relay.
 
      - parameter relay: Target relay for sequence elements.
-     - returns: Disposable object that can be used to unsubscribe the observer from the variable.
+     - returns: Disposable object that can be used to unsubscribe the observer from the relay.
      */
     public func emit(to relay: PublishRelay<E>) -> Disposable {
         return emit(onNext: { e in
@@ -46,10 +68,10 @@ extension SharedSequenceConvertibleType where SharingStrategy == SignalSharingSt
     }
 
     /**
-     Creates new subscription and sends elements to variable.
+     Creates new subscription and sends elements to relay.
 
      - parameter to: Target relay for sequence elements.
-     - returns: Disposable object that can be used to unsubscribe the observer from the variable.
+     - returns: Disposable object that can be used to unsubscribe the observer from the relay.
      */
     public func emit(to relay: PublishRelay<E?>) -> Disposable {
         return emit(onNext: { e in