+++ Tests/RxCocoaTests/Signal+Test.swift
@@ -283,7 +283,48 @@ extension SignalTests {
     }
 }
 
-// MARK: emit variable
+// MARK: emit behavior relay
+
+extension SignalTests {
+    func testEmitBehaviorRelay() {
+        let relay = BehaviorRelay<Int>(value: 0)
+        
+        let subscription = (Signal.just(1) as Signal<Int>).emit(to: relay)
+        
+        XCTAssertEqual(relay.value, 1)
+        subscription.dispose()
+    }
+    
+    func testEmitBehaviorRelay1() {
+        let relay = BehaviorRelay<Int?>(value: 0)
+        
+        let subscription = (Signal.just(1) as Signal<Int>).emit(to: relay)
+        
+        XCTAssertEqual(relay.value, 1)
+        subscription.dispose()
+    }
+    
+    func testEmitBehaviorRelay2() {
+        let relay = BehaviorRelay<Int?>(value: 0)
+        
+        let subscription = (Signal.just(1) as Signal<Int?>).emit(to: relay)
+        
+        XCTAssertEqual(relay.value, 1)
+        subscription.dispose()
+    }
+    
+    func testEmitBehaviorRelay3() {
+        let relay = BehaviorRelay<Int?>(value: 0)
+        
+        // shouldn't cause compile time error
+        let subscription = Signal.just(1).emit(to: relay)
+        
+        XCTAssertEqual(relay.value, 1)
+        subscription.dispose()
+    }
+}
+
+// MARK: Emit to relay
 
 extension SignalTests {
     func testSignalRelay() {
@@ -325,7 +366,7 @@ extension SignalTests {
         XCTAssertEqual(latest, 1)
     }
 
-    func testDriveVariableNoAmbiguity() {
+    func testDriveRelayNoAmbiguity() {
         let relay = PublishRelay<Int?>()
 
         var latest: Int? = nil