+++ Tests/RxSwiftTests/ObservableType+SubscriptionTests.swift
@@ -35,5 +35,31 @@ extension ObservableSubscriptionTest {
         _ = Observable<Int>.error(testError).subscribe()
         XCTAssertEqual(loggedErrors, [testError])
     }
+    
+    func testCustomCaptureSubscriptionCallstack() {
+        var resultCallstack = [String]()
+        
+        Hooks.defaultErrorHandler = { callstack, _ in
+            resultCallstack = callstack
+        }
+        _ = Observable<Int>.error(testError).subscribe()
+        XCTAssertEqual(resultCallstack, [])
+        
+        Hooks.customCaptureSubscriptionCallstack = {
+            return ["frame1"]
+        }
+        _ = Observable<Int>.error(testError).subscribe()
+        XCTAssertEqual(resultCallstack, [])
+        
+        Hooks.recordCallStackOnError = true
+        _ = Observable<Int>.error(testError).subscribe()
+        XCTAssertEqual(resultCallstack, ["frame1"])
+        
+        Hooks.customCaptureSubscriptionCallstack = {
+            return ["frame1", "frame2"]
+        }
+        _ = Observable<Int>.error(testError).subscribe()
+        XCTAssertEqual(resultCallstack, ["frame1", "frame2"])
+    }
 }