+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 * Adds `scan(into:accumulator:)`.
 * Adds `queuePriority` parameter (defaults to `.normal`) to `OperationQueueScheduler`.
 * Performance enhancement reduces Bag dispatch inline code size by 12%.
+* Adds `customCaptureSubscriptionCallstack` hook to allow custom subscription callstacks to be generated.
+* Remove usage of `Variable` from Playground, Example projects and Tests.
 
 #### Anomalies
 
@@ -14,6 +16,7 @@ All notable changes to this project will be documented in this file.
 
 * Adds new `insert` extension to collect and add multiple disposables to `DisposeBag`.
 * Adds `string` to `NSTextView`.
+* Removes string interpolation warning
 
 ## [4.2.0](https://github.com/ReactiveX/RxSwift/releases/tag/4.2.0)